basic separation
